feat(premium): improve files organization
